requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
streamlit>=1.31.0
python-dotenv>=1.0.0
ollama>=0.1.6 